Changed type of inversions to long to be able to handle large arrays

+++ inversioncount.c
@@ -20,9 +20,9 @@ void print_array(char *label, int *array, int size)
     printf("\n");
 }
 
-int merge_and_count(int *array, int left, int middle, int right)
+long merge_and_count(int *array, int left, int middle, int right)
 {
-    int inversions = 0;
+    long inversions = 0;
 
     int i = left;
     int j = middle + 1;
@@ -60,10 +60,10 @@ int merge_and_count(int *array, int left, int middle, int right)
     return inversions;
 }
 
-int merge_sort_and_count(int *array, int left, int right)
+long merge_sort_and_count(int *array, int left, int right)
 {
     if(left < right) {
-        int inversions = 0;
+        long inversions = 0;
         int middle = (left + right) / 2;
         inversions += merge_sort_and_count(array, left, middle);
         inversions += merge_sort_and_count(array, middle + 1, right);
@@ -77,14 +77,14 @@ int merge_sort_and_count(int *array, int left, int right)
 
 int main(int argc, char *argv[])
 {
-    int inversions = 0;
+    long inversions = 0;
     int size = 10;
     int array[] = {9, 8, 7, 6, 5, 5, 4, 3, 2, 1};
 
     print_array("V: ", array, size);
     inversions = merge_sort_and_count(array, 0, size - 1);
     print_array("V: ", array, size);
-    printf("Inversions: %d\n", inversions);
+    printf("Inversions: %ld\n", inversions);
 
     return 0;
 }
